Move prepare lifting pass after convert tpu to cpu pass.
The motivation for moving back the prepare lifting pass after the convert TPU to CPU pass is to avoid unexpected quantization failures due to the bf16 ops handled by the prepare lifting pass such as FusedBatchNormV3. The benefit of locating the pass before the convert TPU to CPU pass was to convert the AddV2 ops to BiasAdd ops in case that the AddV2 ops follow the quantizable ops such as conv2D. At that time, we casted quantizable BF16 ops one by one, but for now we cast all bf16 ops to f32 in the graph. So now the prepare lifting pass will play with all f32 ops.

PiperOrigin-RevId: 517287798