Throw error in TFL->Tosa lowering for grouped convolution.
TFL supports grouped convolution when kernel channels is a factor of input channels, but Tosa doesn't. As a result, the lowering generates an invalid shape and tries to convolve it as normal, but fails on shape validation in tosa. Grouped convolution lowering for tosa WIP, but in the meantime we should throw an error for this specific case.

PiperOrigin-RevId: 517470238