Fix undefined behavior in saturate_cast.
Our clip+cast approach for `saturate_cast` can lead to undefined behavior in cases
where the output range is not exactly representable in the input type.  For
example, in `float32` to `uint32` casts, the upper `uint32` bound `4294967295` cannot be
represented in `float32`, which instead uses `4294967296`.  After the `clip` operation,
we may then have values of `4294967296` in the input tensor, outside the
valid range of `uint32`, and results in UB when cast (and odd results).

To fix this, we need to nudge the clip limits to be within the valid bounds of the output
type.  We do this via `nextafter()` when necessary.  A consequence is that we don't
actually necessarily clip to the largest value - e.g. `inf` will get saturate-cast to
`4294967040` instead of `4294967295` - however, we can't fix this without introducing
a new `SaturateCast` op that knows about both the input and output types.

PiperOrigin-RevId: 517505444