#tf-data Fix potential check failure in autotune recording.
The issue was that autotune recording was being turned on accidentally during parallel interleave's save method, which called GetNext on its input. Previously GetNext assumed that a downstream iterator would always have recording enabled before calling GetNext, so GetNext would always "turn it back on" at the end.

This CL updates GetNext to preserve the original recording state, whether it was on or off.

This CL also makes a similar change to RunWithBorrowedArgs, which previously made the same assumption.

PiperOrigin-RevId: 517537441