Add an option to toggle running `RemoveVariablesInSessionInitializerPass`.
* **Preserves default behavior except for the TF Quantization passes.**

This change adds a new field to `MLIRImportOptions` so that users can enable / disable running `RemoveVariablesInSessionInitializerPass`.
Variables in session initializer functions may not be removed, especially when variables are not lifted, as they are not hoisted as global variables and variable initialization patterms should be preserved.
This applies especially when variables are used to initialize other resources (like hash tables).
Removing variables will remove the hash table initialization pattern altogether, which may result in the "table not initialized" error.

The default value of `MLIRImportOptions::include_variables_in_initializers` is set to `false` to prevent breaking existing usages.

They are set to `true` for TF Quantizer passes.

PiperOrigin-RevId: 517947036