*** Reason for rollback ***
The mitigation can be removed after the proper fix to the issue is submitted.

*** Original change description ***

Add timestamp to the accuracy benchmarking storage buffer path.

It allows different accuracy benchmarking runs to have an unique name. It prevents the file lock collisions between the previous benchmark runs and a new benchmark run.

PiperOrigin-RevId: 518305664