WeakRefCounted sends notification before destructors.
This makes the behavior inside the weak notification more defined.
e.g. if the weak notification's receiver is owned by the WeakRefCounted that sends the notification, then after this change the receiver is guaranteed to be alive when the notification is sent. Before this change it is guaranteed to be dead, which is not very convenient, and actually causes failures in the child CLs regarding the rendezvous leak fixes by using WeakPtrs.

PiperOrigin-RevId: 518311417