Convert local rendezvous table to a Rendezvous creation cache using WeakPtr.
A new EagerContext stat is added for inspecting number of 'leaked' local Rendezvous objects. I checked this metric manually and with this CL there is no apparently leak from a full train step from save_restore_v2_test.py.

When a rendezvous has no pending requests it will be automatically destroyed.
This is achieved by tracking a reference count of the owner rendezvous from the Item member of the LocalRendezvous. Clamping at where the rendezvous is used is conceptually simpler than dealing with the 4 special cases individually.

The global rendezvous is kept alive by the eager context.

PiperOrigin-RevId: 518647764