Fix memory management bug in unit test.
The TfLiteOpaqueContextReplaceNodeSubsetsWithDelegateKernels
API function is documented as taking ownership of the
TfLiteRegistrationExternal* parameter: see
https://github.com/tensorflow/tensorflow/blob/757d6a0db6d11a8fb78984544882f036a733f847/tensorflow/lite/core/c/c_api_opaque.h#L265
But the code in this test was storing the TfLiteRegistrationExternal pointer
in the DelegateState struct, and the passing it to
TfLiteOpaqueContextReplaceNodeSubsetsWithDelegateKernels in the Prepare
method.  Since the Prepare method can in general be called multiple times,
and since the TfLiteRegistrationExternal* field in the DelegateState
struct could become a dangling pointer after the first call to Prepare,
this isn't safe.  Instead, the test should create the TfLiteRegistrationExternal
object each time that TfLiteOpaqueContextReplaceNodeSubsetsWithDelegateKernels
is called, in order to guarantee that it is being called with a valid object
whose ownership can be transfered.

PiperOrigin-RevId: 518824383