[XLA:GPU] Fix cudnn_vectorize_convolutions when the input/output don't have the same dnums.
The code incorrectly beleived that conv side input should use `input`'s dnums.
In fact it should use `output`'s dnums; the side-input is added *after* the
conv is computed.

This works so long as input and output have the same dnums, which is common.
But if you pass a conv with different dnums for inputs/outputs, we eventually
hit a CHECK-failure.

PiperOrigin-RevId: 518866657