Check that zero point and scale of inputs and outputs match for Concatenate operator.
The initial checks within VisitConcatenationNode all pass as xnn_define_concatenate is not called when checking if a node may be delegated or not. It is during the second pass that the check fails, causing delegation to fail.

INFO: Created TensorFlow Lite XNNPACK delegate for CPU.
ERROR: failed to delegate CONCATENATION node #31
ERROR: failed to delegate CONCATENATION node #36
ERROR: Node number 55 (TfLiteXNNPackDelegate) failed to prepare.
PiperOrigin-RevId: 518874585