Move resource_variable_ops proto functions and their registrations to variable_scope.py.
This is in preparation for moving variables.VariableV1 and variables.RefVariable to their own files (in order to remove dependency cycles), which entails moving the variables.Variable.from_proto implementation to variables.VariableV1.

resource_variable_ops.py cannot introduce a dependency on VariableV1, or the existing import cycles will be perpetuated, so moving the from_proto function to variable_scope makes sense as it must already import VariableV1 and ResourceVariable by definition.

PiperOrigin-RevId: 519062300