Reserve function name after signature_wrapper
The function name will be used in graph generation as the function name. In current implementation, all functions in the graph will be named like `__inference_signature_wrapper_xxx` instead of the original function names. This change fixes it.

It may be beneficial to use `tf_decorator.make_decorator`. But that also sets __wraps__ and caused other errors.

PiperOrigin-RevId: 519201710