Refactor common_runtime int32 placement (and optionally log full type mismatch)
Refactor int32 specific placement mechanisms in core/common_runtime that decide if a tensor is on host (and not device) using the helper functions in `arg_ret_placement.*`. In this CL, the decision is made in the same way as previously using dtype, so there is no behavior change. Full type information with `TFT_SHAPE_TENSOR` to indicate host tensors (created by `int32_fulltype.*`) is compared with the decision based on dtype and any mismatches can be logged by using
`vmodule=arg_ret_placment=1` (i.e. the `Weak*` methods from `arg_ret_placement.h` are used).

PiperOrigin-RevId: 519833471