Make Uniquefy algorithm stable in TFLite GraphInfo tool.
TFLite Delegate interface doesn't really guarantee the order of
inputs / outputs. This is for working around some issues that
relies on the order of the inputs and outputs. However the behavior
should not be relied on, in theory.

The code path should be already well covered by the existing tests.
As described above, flipping the inputs/outputs orders in the test
cases should not affect anything.

PiperOrigin-RevId: 519916538