Make tfcompile more friendly on other archs.
Currently tfcompile and tfcompile_wrapper use hard coded values for target_triple. The hard coded value is `x86_64-pc-linux`. This makes the following test(s) fail on s390x:

//tensorflow/python/tools:saved_model_cli_test

I've removed the hard coded string and changed it so that the default value from llvm is used instead.

In the case where a specific target_triple is desired such as in the `testAOTCompileCPUFreezesAndCompilesVariablesToFeedNoneTargetAarch64Linux` and `testAOTCompileCPUFreezesAndCompilesVariablesToFeedNoneTargetAarch64Android` subtests the set value is still allowed to follow through.

The subtests above cause an additional failures on big endian systems due to a check for little endianness. I added an exception to skip these two subtests on s390x.

tfcompile_wrapper doesn't set a default target_cpu. This is causing the following test case to fail on s390x:

//tensorflow/python/tools:aot_compiled_test

I've made a similar change where I use the value from llvm as the default.