common_runtime int32 placement matches full type information
int32 specific placement mechanisms in core/common_runtime that decide if a tensor is on host (and not device) compare full type information with `TFT_SHAPE_TENSOR` to indicate host tensors (created by `int32_fulltype.*`) with the decision based on dtype and raise an error if the decision would be different. (Otherwise, there is no behavior change.)

Potentially, errors could be caused by either by bugs in creating full type information or in graph rewrites that change node outputs without making corresponding changes to full type informations (for cases that are not covered by current tests). If these occur and the root cause is in TensorFlow, please rollback/file a bug and provide a reproducer.

PiperOrigin-RevId: 520187579