Update debug tests to use object-identity dicts for tensors.
A planned change in TensorFlow will remove one of the
existing exceptions to the disabling of `Tensor.__hash__`.
This breaks these debug libraries.

This change updates these libraries to use the existing
`ObjectIdentityDictionary` type from the `util` package.
These dictionaries provide the same functionality as before
but are explicit about their use of reference equality for keys.

PiperOrigin-RevId: 520353720