[tf.data] Update symbolic checkpointing implementation to support both `WriteScalar(name, key, value)` (and `ReadScalar(name, key, *value)`) in addition to `WriteScalar(key, value)` (and `ReadScalar(key, *value)`).
Supporting the "name + key" version of the Read/Write methods is necessary to being able to checkpoint iterator state whose individual parts are each less than 2GB but its total size is more than 2GB. The "key only" version of the Read/Write methods serializes all state of the same iterator using a single proto, while the "name + key" version of the methods makes it possible to use separate protos.

Note that it is not expected for symbolic checkpoints to be large but to avoid divergence between symbolic and explicit checkpointing logic, we want to use the same serialization and deserialization logic for both.

PiperOrigin-RevId: 520469126