Add a nested eager executor, and wire single device Ops through it.
This is a very simple implementation to test the waters.

Some rewiring of small constant folding and const eager Op is done to align it better with how we may run them in the future.

I am trying to stick with the C++ API, since we don't really have a good C API for EagerExecutor. But this introduced some strange boiler-plate code, which I don't know if is the best practice.

PiperOrigin-RevId: 521012217