[DelegatePerformance] Export the path of the folder containing the stable delegate shared library to the STABLE_DELEGATE_LIBRARY_PATH environment variable.
When downstream shared objects are not specified while compiling the upstream shared objects, the stable delegate library is responsible for loading the downstream libraries via dlopen for all necessary symbols. As such, the stable delegate library must have the absolute path to dlopen the downstream shared objects. This change adds an environment variable to better address this case, which contains the stable delegate shared library folder path as value. This will allow the stable delegate library to rely on the environment variable in forming absolute paths.

PiperOrigin-RevId: 521753846