[Linaro:ARM_CI] Stop using ambe config as not needed
The use of the ambe config to build and test aarch64 is
not needed and so stop using it. The ambe config will be
removed in the future so this must be done in preparation.
Also make cpu_arm64_pip.sh and cpu_arm64_nonpip.sh more
similar for easier future maintenance.