update padding='SAME' in tf.image.ssim API
Currently the API tf.image.ssim is using padding='VALID' internally. But with this padding the shape of the output is not matching with what documentation states. With return_index_map=True the output shape should be broadcast(img1.shape[:-1], img2.shape[:-1]). But with padding='VALID', for input of shape     
(16, 2106, 80, 1) the current output shape is (16, 2096, 70) which is not matching with the expected output which should be (16, 2106, 80). With padding='SAME' the output is matching to the desired output. 

Please refer to the attached gist showcasing all the mentioned details.
https://colab.research.google.com/gist/SuryanarayanaY/d75ef4fcbd69b5211975fd4d29490084/59067_r1.ipynb