Remove some unnecessary assumptions from DotMerger pass.
It doesn't need to assume that DotDecomposer has run before it. What it really
needs is that there is just a single non-contracting dimension in the two dot
operands to be concatenated, as concatenate only supports concatenating along a
single dimension.
Also, add some checks in case the pass is run after layout assignment. We need
that the layout of the two dot operands to be concatenated is the same, and the
layout of the two dots to be merged is the same.
We don't expect any changes to generated code, as currently the assumptions
which are removed still hold.

PiperOrigin-RevId: 522309291