Remove mesh_ from TensorWithLayout.
This member is redundant to layout().mesh().

This change also eliminates the dangerous use of const Mesh& in TensorWithLayoutTf.
It may silently refer to destroyed object instances, either from stack or
from heap. If we want to reduce the memory footprint due to having too many
Mesh objects persisted, then we shall upgrade Layout::mesh_ to a pointer, or
even a ref counted pointer.

This change builds on top of the diffbase.

PiperOrigin-RevId: 522365961