Handle `XlaGatherOp` -> `SliceOp` conversion when the output shape's rank is known.
The shape of the output of `XlaGatherOp` is a result of collapsing the `collapsed_slice_dims`.
When decomposing `XlaGatherOp` to `SliceOp`, etc., the shape of the output needs to be expanded back in order to determine the output shape of the `SliceOp`.
The subsequent `ReshapeOp` will collapse the `collapsed_slice_dims` back, matching the original output shape of `XlaGatherOp`.

See https://www.tensorflow.org/xla/operation_semantics#gather for detailed information of the semantics of `XlaGatherOp`.

PiperOrigin-RevId: 522531148