Make LLVM patch more robust
Instead of reverting the entire commit 1ef51e0452a4, which introduces a problematic dependency edge, we just exclude the files from @llvm-project//mlir:Analysis.

This was causing problems for TensorFlow as described in
https://github.com/tensorflow/tensorflow/commit/052a3663f7f2cb06a65a926e0b3684e7a2b06c94

If in the future we need to depend on them and the linking issue is not yet solve, we can just create a separate target for them.

PiperOrigin-RevId: 522868341