Run single device cluster through the MLIR.
Interestingly that the current Mesh propagation and layout propagation logic
'worked' as is for the very simple test case we had.

We should probably clean up and refactor the 'launch' section of ExecuteRegularOperation to make it more readable, but I wonder if that investment is good pay off as DTensorNG will bypass the entire thing. Perhaps we shall do that clean up to just learn of all of the tricks that are involved there.

PiperOrigin-RevId: 523162723