Do not render "Current Behaviour" box as shell in issue template.
The GitHub issue template currently renders the "Current Behavior" box as a shell. This makes it difficult to read, as the text is monospace and doesn't wrap, instead showing a horizontal scroll bar if the line is too long.

See https://github.com/tensorflow/tensorflow/issues/59281 for an example of how rending as a shell makes the "Current Behavior" box difficult to read.

PiperOrigin-RevId: 523508635