Don't reassign distributed snapshot streams.
Supporting stream reassignment without lease files potentially allows a race condition among multiple workers contending on the same stream.  This interim change establishes a 1:1 relationship between worker and stream.  Note that as of this change the number of workers in a single snapshot process should only ever increase, as if the number of workers decreases, the departed workers' streams will never be reassigned and their uncommitted chunks will be lost.

PiperOrigin-RevId: 523536141