Don't clean up distributed snapshot workers.
Before this change, the failure mode was a worker dying and not returning during the snapshotting process and its uncommitted data being lost.  With this change, the failure mode is a worker dying and not returning and the snapshotting process never terminating.  The latter is preferable because it guarantees exactly-once visitation.

PiperOrigin-RevId: 523788640