[tosa] Add TFLite Python flatbuffer conversion entry point.
Create a convenience wrapper that does the conversion from TFLite flatbuffer to
MLIR bytecode with TOSA ops for consumption by compilation backend. This is a
more constrained and stabler entry point than the previously exposed
pass-pipeline string usage (it also avoids many of the conversions back and
forth as strings).

PiperOrigin-RevId: 524415140