Divert calls to MklMatMulOp into BatchMatMulMkl for aarch64
We observed that for most transformer models, the MklLMatMulOp
node in tensorflow gets called. This operation requires an SGEMM
BLAS interface which is not supported by the Arm Compute Library
(ACL). Hence, when running with TF_ENABLE_ONEDNN_OPTS=1, the
sub-optimal gemm_api kernels are used in oneDNN (instead of
the ACL ones). Here, we divert calls to MklMatMulOp into
BatchMatMulMkl for aarch64 which allows the ACL kernels to be used
when running with TF_ENABLE_ONEDNN_OPTS=1.