Support GELU as an elementwise operator in the TFLite GPU delegate.
GELU activations are commonly used in Transformer architectures. The gelu op specification allows for a choice between an approximate and accurate computation.
The accurate version involves the use of erfc, which is not present in OpenGL or Vulkan.

For simplicity, the OpenCL version always uses the accurate form and OpenGL and Vulkan always use the approximate form, ignoring the value set in the op attributes.

PiperOrigin-RevId: 525660188