[XLA:GPU][NFC] Make CanFoldTransposeIntoDot handle non-canonical dots.
If there is more than one non-contracting dimension, and the layout of the
transpose operand still makes sure that the non-contracting dimensions are
consecutive, it needs to be checked whether the dot shape is still the same after
folding.
Copy several end-2-end tests of folding transposes from gemm_rewrite_test and
make them unit tests, and add one test covering the additional logic.
This change is NFC for now, because we only run TransposeFolding after DotDecomposer, so there currently is always only one non-contracting dimension.

PiperOrigin-RevId: 525661973