During GPU layout assignment, don't assume dots have been canonicalized.
This is a first step towards potentially moving DotDecomposer to after layout
assignment.
Also rename some variables to make it clearer what is going on. We were swapping
contracting and non-contracting dimensions twice for the lhs of a dot, so simplify
that to avoid swapping twice.
No impact expected on benchmarks, as we currently do run DotDecomposer before
Layout assignment.

PiperOrigin-RevId: 525672122