Fix build breakage on iOS for gpu_plugin-related targets //tensorflow/lite/experimental/acceleration/configuration:gpu_plugin_impl //tensorflow/lite/core/acceleration/configuration/c:gpu_plugin caused by commit 9765df1a06107e3dae851ce38259d25767573058.
There were two issues.  The first one is that we need to ensure that
the gpu_plugin.h header file is built as Objective-C++, not as C++; this is
important because Bazel compiles the header files to verify that they are
self-contained, and this header (conditinally) includes the metal delegate,
which references Objective-C headers, so it won't parse as C++.

The other issue is that it was missing a conditional dependency on the target
for the metal delegate.

PiperOrigin-RevId: 525711555