Propagate errors from async checkpoint's async thread to the calller thread.
One of the current issue with TF2 async checkpoint is that errors happening in the async thread are silently ignored as they do not reach the main thread. This CL resolves this issues by propagating the async thread error to the main thread so get the caller's attention.

The implementation stores the async thread error as the internal state in the `async_checkpoint_helper` and emerges it the next time async save/write is called, or when the async save thread is joined.

PiperOrigin-RevId: 525875359