Stop caching list of devices in WorkerEnv
This change to prepare for 'dynamic' virtual device support in tensorflow single controller
system's workers, which will make DTensor and outside compilation in general
easier on these workers.

In that case, the worker's virtual device list can be mutated via a WorkerService RPC call (see comment in cl/525889023), making it more consistent with the Python execution path. Then the list is no longer static and the caching gets into the way.

Even in the current status, it is unclear which of the users actually benefit from having the cache. (none of the users appear to be on a performance critical path).

I am also renaming a few StaticDeviceMgr to DynamicDeviceMgr, as the former has been converted to an alias of the later earlier of the day.

PiperOrigin-RevId: 526082955