Add specialized GPU TopK kernel.
This faster version works for K<16 and relies on keeping a sorted list of values
in registers to avoid data movement as much as possible.

The algorithm is essentially broken down into two sections:
 1. Slice the data equally betwen warps and compute the warp-local topk.
	- First we split the data into equal-sized chunks, load and sort the first K
		elements in registers.
	- Next each thread iterates over the remainder of its data, keeping the
		in-register array sorted.
	- Finally, on each warp we reduce the lane-local topk to a warp-local topk.
	- One thread for each warp writes the warp-local topk to shared memory.
 2. On a single warp:
 	- Lane i reads the warp-local topk from warp i.
	- Now that each lane has the corresponding warp-local topk, compute the global
		topk.

For efficiency, we pack float32 & bf16 into u64/u32 togehter with their indices.
This promotes better codegen and cheaper comparisons.

PiperOrigin-RevId: 526185694