[XLA] Fix latency hiding scheduler delay send-done logic.
The send done logic tries to delay send-dones to avoid triggering them too early.
This logic was broken, because the ready time of the send() attached to the send-done()
is not actually populated as it is not ready yet. We need to estimate as the ready time
of the successors + latency.

This was causing the delay logic to always triggering making effectively kScheduleDone policy
never trigger for send-dones() host.

PiperOrigin-RevId: 526200252