Fix a bug in allocation_test.cc where it was not testing what it was intended to test.
The test case named "allocation_excessive_length" was
a cut-and-paste duplicate of the "allocation_invalid_length"
test case immediately above it, and so was testing an invalid length (zero)
rather than an excessive length.  There was no test for an excessive
length.  The fix was to change the length parameter in that test case
from zero to the minimum excessive value (file length + 1).

PiperOrigin-RevId: 526991265