Clean up and move around helper constraints so we can use them in multiple files.
Move `HasSameStaticShape` constraints from `legalize_patters.td` to `utils.td` so that they can be used in multiple files. Rename to `OpHassameStaticShapes` to better reflect the behavior of this constraint.

Also delete `HasOneUse` from `utils.td`. There are many duplicates of this constraint in various included files (`tensorflow/optimize.td` for example). This was causing an issue where `utils.td` could not be included. `HasOneUse` was only used in `optimize_patterns.td` so we just add it in that file instead.

PiperOrigin-RevId: 527128318