Change sparse_to_dense tests to avoid writing to the model data.
When the test type is 'TestType::kConstant', the 'output_shape' input
parameter to the model is already set to a constant when the model is
constructed, and there's no need to re-set it to the same value again.

Without this change, these tests would fail if the model was stored
in read-only memory.

PiperOrigin-RevId: 527391309