Fix TF32 test and docstring.
Even when TF32 is enabled, cuBLAS is not guaranteed to use TF32 and it is not documented in which cases TF32 will be used. Before, the TF32 test used matrix shapes of (8, 8) but this currently isn't causing TF32 to be used on an A6000 GPU, although it does use TF32 on an H100. Similarly, the docstring used matrix shapes of (2, 2), causing TF32 not to be used on an A6000. I changed the shapes to (1024, 1024), which does use TF32 on an A6000 (and an H100) as of CUDA 11.8.

PiperOrigin-RevId: 527932808