Fix TopKDecomposer stability.
When I wrote this code I expected the TopKDecomposer to run after the
StableSortExpander, but reality this needs to run earlier in the pipeline. This
means the to_apply comparator used for the TopK doesn't take indices into
consideration and we need to expand the sort into a stable sort.

PiperOrigin-RevId: 529154170