Rename RelayoutGrad to RelayoutLike
This is toward the proposed Op names in the RFC.
https://github.com/tensorflow/community/pull/442

In addition: RelayoutLike can be a tool for implementing tf.ones_like.
The python tf.ones_like implementation doesn't use tf.OnesLikeOp.
It is unclear if we have to avoid tf.OnesLikeOp; if so, then we shall emit a RelayoutLikeOp in graph mode after
the array creation call.

PiperOrigin-RevId: 529524377