Fix negative comparison order.
The previous version didn't flip non-sign bit, resulting in the wrong
order for negative values (e.g. 200, 100, -200, -100 instead of 200, 100, -100, -200). This wasn't caught in tests because when there's a mix of positive and integer values, the topk elements are positive with extremely high probability.

Removing the bitpacking results in a 20% performance loss which we consider acceptable for now, given this kernel provides a 2-10x speedup over the sort+slice version (depending on K and batch size).

PiperOrigin-RevId: 529756895