Set XLA PrecisionConfig in tf2xla bridge based on whether TF32 is enabled.
The net effect of this change is that on TPUs, disabling TF32 will now cause dots/convs to run in FP32 precision instead of BF16, as the PrecisionConfig is set to HIGHEST instead of DEFAULT.

Currently this does nothing on GPUs as XLA directly checks if the TF32 global variable is set. But soon XLA will stop using the global variable and instead check the PrecisionConfig, since it's bad for XLA to rely on TensorFlow global variables. The primary purpose of this change is to make removing XLA's use of the global variable possible.

PiperOrigin-RevId: 530378023