Fix an issue where MlirBridgePass::GetPassState may incorrectly count TPU graphs as CPU/GPU graphs.
Currently GetPassState is called twice, first before Run, and then during if the first passes. GetPassState incorrectly assumes all graphs are for CPU/GPU when called during Run. This causes a TPU graph to incorrectly skip graph ananlysis, as graph analysis is not done for CPU/GPU graphs. This may causes issues if such a graph get transformed between the first and second GetPassState call from a valid graph to an invalid one. The graph will continue to run on the MLIR bridge in this case, even if it should not. This fixes the issue.

PiperOrigin-RevId: 530454825