#tf-data Export NumpyIterator type.
The type is already returned from the `Dataset.as_numpy_iterator()` API. Now users will be able to refer to the type, which is useful e.g. for registering a serializer with Flax.

This CL also removes the underscore before the _save and _restore methods, since they are stable and intended to be used outside of TF.

PiperOrigin-RevId: 530539643