Support tf.summary.image in DTensor.
An SPMD expander is added to support the vectorized EncodePng Op.

Note that we are lucky that Layout propagation makes the EncodePngOp itself fully replicated, emitting a collective on the int8 input, rather than emitting a collective reduce on the string output. Currently we only support int8 collectives, not string collectives.

PiperOrigin-RevId: 530706098