Delete PJRT DeviceCompiler (and cache) if PjRtClient is stale
We check whether the PjRtClient has changed or not before compilation. We delete and create a new PJRT DeviceCompiler if it has changed.

Ideally the cache should be invalidated upon shutdown, although getting a hold of the TpuDevice in TpuSystemShutdown() doesn't seem straightforward (the cache is stored in the TpuDevice's resource manager).

PiperOrigin-RevId: 530751696