Fix incompatibility between LazyLoader and doctests.
In certain tests this fails at
the line `for valname, val in obj.__dict__.items()`,

https://github.com/python/cpython/blob/22f3425c3d3d896be0917d80d55e8abb08d99b18/Lib/doctest.py#L1014

because LazyLoader may mutate __dict__ when certain `__` attributes of val is accessed. We will have to come back to update this when doctest changes. The added unit test will hopefully help us to catch these problems in the future.

PiperOrigin-RevId: 530967131