Stabilize the gradient for tf.math.reduce_std.
As the standard deviation of a tensor approaches zero,
the gradient unfortunately becomes undefined because of
a singularity of `dsqrt(v)/dv` at zero. However, we know that
the gradient of variance is always bounded.  Even though
it is not technically defined at exactly zero (it approaches
different values depending on the direction in which we
take the limit), we return the average, which is zero.

This is to fix a common use-case where data is normalized
to be distributed `~ N(mean=0,std=1)`.

PiperOrigin-RevId: 531233567