Update Args data types of tf.ones and tf.zeros
At present as per documentation of tf.zeros and tf.ones, both APIs have argument shape which accepts "A list of integers, a tuple of integers, or a 1-D Tensor of type int32 ". But its find out that when Boolean data types True, False passed to the same argument it also accepts it and convert them into 1,0 respectively and outputs correct results. Hence i am proposing to add a note under  the 'shape' argument description that it also accepts boolean data types.

Attaching the gist also for referring the results.
https://colab.research.google.com/gist/SuryanarayanaY/837bf1888c618e3585f4f7247c329885/tf-ones_tf-zeros_tf-eye.ipynb