[XLA:GPU] Handle non-standard output layouts in MakeDotSplitKBatch
The output layout was not set previously, resulting in TritonAutotuner discarding split_k > 1 configurations when a non-standard output layout was specified in the HLO.

This error was unlikely to cause problems for users, because it was only triggered with non-standard output layouts and even if was triggered, the TritonAutotuner likely rejected the incorrect configuration.

The implementation of the test `SkipConfigsProducingDeviantResults` is not applicable anymore, because the result is not deviant after fixing this corner-case.

I may upload a new `SkipConfigsProducingDeviantResults` test case in a later cl, after carefully crafting a deviant example.

PiperOrigin-RevId: 532492934