Enforce that the actual arguments for XlaCallModule match the expected types and shapes.
Previously, when we loaded the MLIR module in preparation for shape refinement we override the argument types with the types of the actual arguments, without checking first if this is safe. If the actual arguments do not have matching types we may run into errors downstream (during shape refinement, or during the compilation), but in some cases there may be silent errors.

Here we add explicit checking that the types match.

PiperOrigin-RevId: 532968976