#tf-data-service Use total split count to track progress.
This (1) fixes a bug in cl/532322377 where splits were only considered to be completed at stream completion time, (2) abstracts away the notion of streams which should be considered an implementation detail, and (3) provides a better view of the whole progress by showing the total number of the splits.

Note that this does *not* handle repetitions which are generally supported as of cl/532571816.

PiperOrigin-RevId: 533200133