Fix link error in transpose_op.cc.
The template CSRSparseMatrixTranspose, was used in a file where the template definition was unavailable. This requires the template to be explicitly instantiated, but it wasn't explicitly instantiated before.

This issue has been around for a long time. I'm not sure why this wasn't failing before, but it's now failing for me.

PiperOrigin-RevId: 533344851