[XLA:GPU][NFC] Use AddEntryComputationWithLayouts in AlgebraicSimplifierTest
AddEntryComputationWithLayouts:
Same as the AddEntryComputation function but the module's
entry_computation_layout is updated to match the layout of the new entry
computation.

This is needed for a future CL where we will add stricter compatibility checks between the entry_computation_layout and the entry computation's input/output layouts.

We're submitting this in a separate CL because these changes may "hide" more important changes in the coming CL.

PiperOrigin-RevId: 534033498