Fix undeclared inclusions in CUDA build with GCC on symlinked path
When the host compiler (e.g. GCC) is installed in a path
being/containing a symlink the CUDA build might fail with `"undeclared inclusion(s)"`
to the symlink paths.

This has been partially solved by #56360 which resolves all symlinks in `gcc_host_compiler_path`
However the compiler referenced there might be a wrapper script (e.g. ccache)
which internally eventually forwards the call to the real compiler
sitting in a path containing a symlink.
`gcc -v -E - < /dev/null` seems to only return resolved paths
while the later invocations include the paths with symlinks leading to the build failure.

The solution employed here is to add the unresolved paths (i.e. the ones
with symlinks) in addition to the resolved paths.

Fixes #33975