Do not partially quantize the inputs of UnidirectionalSequenceLSTMOp
The TFLite runtime doesn't currently support hybrid/partial quantization of the
inputs for the UnidirectionalSequenceLSTMOp. As a consequence, we must ensure
quantization occurs atomically for all of the input constants. In cases where
any of the input constants do not meet the minimum element count for
quantization eligibility (as defined by the `minimum_elements_for_weights`
quantization config parameter) then skip quantization for this op entirely.

PiperOrigin-RevId: 534562296