#tf-data Reduce state when symbolically checkpointing flat_map.
Previously we would store the input element that created the current nested dataset, so that we could re-run the flat_map function to regenerate the nested dataset.

With this change, we instead replay the latest input to flat_map to recreate the input element for the nested dataset. This way, flat_map doesn't need to checkpoint any (potentially large) tensors.

PiperOrigin-RevId: 534958962