Create a super flag for using PJRT for single device compilation + execution.
This enables the feature regardless of the device type and the other op-specific flags being set or not.

This is mostly for the ease of developing Python E2E tests. Currently, we have to 'allow' specific device types programmatically in-addition to setting the env var flags. This works well for C++ unit tests, but not so much for Python tests.

PiperOrigin-RevId: 535323103