Change TF custom calls in `XlaCallModule` op to use indexes in the `function_list` to refer to TF functions
The previous implementation based on embedding TF function names as string turned out to be too brittle because various TF-level graph rewrites may rename functions without knowing that the names of those functions are embedded in XlaCallModule's serialized StableHLO.

Instead, this CL uses a `called_index`, which stores an index in the `function_list` of `XlaCallModule`. Since TF graph rewrites that rename functions will rename function names in `function_list`, this representation ensures that function names are always in sync.

PiperOrigin-RevId: 535426269