Lift HashTable ops as function arguments
In the graph execution mode, resource ops with the same `shared_name` attribute point to the same underlying resource. This is not true in the eager execution mode. This cl lifts HashTable ops as arguments, updates its caller and unifies the lifted ops with any existing in the caller function.

PiperOrigin-RevId: 535495431