Record and show the erroring clusters in Mesh propagation mismatch.
1. This CL records the originating argument index as the dtensor.from_arg_index attribute on the DTensorLayout ops created in PropagateDefaultLayout. The reason
this attribute is kept on the DTensorLayout ops are 2 folds:
  a) DTensorLayout ops is added as a signal that the Op is created from inputs.
     since nodes with layout attrs are almost always from inputs.
  b) ConstOp nodes are ephemeral. They die and reborn during SCCP lossing the
     attributes accumulated during the previous lives.

2) When Mesh propagation fails with inconsistent meshes, this CL ensures the
error message shows all of the inputs have been scanned this far.
Making this information available eliminates the need during debugging
to rerun with flags that produce graph dumps, which can be cubersome (can
take days in a coordinated multi-eng cross team colaboration workflow).

After this CL, the error message contains back references to the graph argument
index if any of the erroring clusters are from graph arguments. This helps
future steps in tracking down the origin of the erroring input to the graph.

PiperOrigin-RevId: 535698991