Add a test that fails when the func renaming logic in XlaCallModule lowering doesn't work correctly
This is to increase the coverage of XlaCallModule's custom call lowering, which involves renaming functions to avoid naming collision. The test is carefully constructed such that there are two custom calls with the same name but different body computations, so that the test fails if the renaming logic doesn't work correctly.

PiperOrigin-RevId: 535741974