validate tf.experimental.numpy.stack axis argument
AT present tf.experimental.numpy.stack not validating the axis argument. If we pass out of bound values to the axis still it is not rising the Error. Hence I am proposing amendment to validate the axis argument and raise the Error in case of out of bound axis.

Please find the attached gist for explaining the problem and also with proposed amendments to the code it works fine and raised intended error when invalid axis arguments provided.