Make it possible to call TF functions that take tokens from a custom call op that does not
Currently, whether the inner TF function is lowered to have a token argument/result is decided based on whether the custom call op takes a token or not. This makes it impossible to call inner TF computation that requires a token (via `_xla_token_input_nodes`) from a custom call that does not take a token.

Instead, this makes `XlaCallModule` behave more like functional control flow ops, where the existence of non-empty `_xla_token_input_nodes` attributes (i.e., whether the op itself takes a token or not) determines whether the inner computation is lowered with a token argument/result. If the custom call is missing a token but the inner computation is lowered with a token argument/result, a dummy token is created and used as the token argument during invocation.

PiperOrigin-RevId: 536458490