[xla:cpu:next] runtime: split convolution target in two
One target depends on XLA Runtime's custom_call, while
the other one doesn't. This will soon allow us to call
the latter in the AOT flow, which cannot depend on
custom_call.

While at it, rename xla.cpu.convolution to xla_cpu_convolution
so that we can define it as a C standalone function.

PiperOrigin-RevId: 537322116