Do not sort the broadcast dimensions during layout normalization.
The order of broadcast dimensions is semantically meaningful, as it allows to express a transpose as part of broadcasting. This is usually canonicalized by the BroadcastCanonicalizer pass into a broadcast with sorted dimensions + transpose. However, if dimensions are unsorted when they are fed to normalization, they have to remain as such. This is ensured by this change.

PiperOrigin-RevId: 537893427