Break the interdependency between tpu/tpu_strategy_util.py,  distribute/cluster_resolver/tpu/tpu_cluster_resolver.py, and eager/remote.py.
Make exported wrappers for initialize_tpu_system and shutdown_tpu_system in tpu_cluster_resolver.py, and make implementation functions for them in tpu_strategy_util.py that they pass a reference of TPUClusterResolver to. Add a test for this.

Replace an isinstance check of TPUClusterResolver in remote.py with an isinstance check of ClusterResolver and a hasattr check.

Make all inline imports regular imports and add all missing BUILD deps. Update references to the new location of initialize_tpu_system.

PiperOrigin-RevId: 537905952