tf.size supports TF_FLAG_TF_SHAPE_DEFAULT_INT64=true
Use environment variable `TF_FLAG_TF_SHAPE_DEFAULT_INT64=true` to cause the output of tf.size to be int64 if `out_type` is not specified (and the flag similarly affects tf.shape). (See example in the BUILD file of how to do this for a test.) Setting this flag to true is highly experimental / not yet supported / causes known breakages.

PiperOrigin-RevId: 538253301