Add disabled_checks attribute to XlaCallModule.
This attribute encodes the disabled_checks option that was passed at
serialization time, and also controls safety checks when the module
is loaded and compiled. See docstring in xla_ops.cc.

Also bump the version number of XlaCallModule to 6.

From this version the XlaCallModule must have a non-empty `platforms`
attribute to encode the platforms for which the module was serialized.
Previously the front-ends could disable the platform checking
by setting an empty `platforms`.

We have not yet added a way to add more disabled checks using an environment variable.

PiperOrigin-RevId: 538421669