Fix bug to allow shapes to be passed explicitly to per-worker variable, and add to the docstring on usage and current caveats. Add a test for explicitly passing shapes.
The issue here was that the placeholder tensor used to represent the per-worker variable during tracing did not have a resource handle attached to it. And any placeholder tensor that does not represent a resource has its shape clobbered, so "unknown" shapes were used, which some ops (like tf.boolean_mask) don't like. The solution is to attach a handle to the placeholder tensor using a custom spec to let the shape be used during tracing.

PiperOrigin-RevId: 538604734